Create folder for protocol contexts
